\name{Chocolate}
\docType{data}
\alias{Chocolate}
\title{Chocolate production and quality measures}
\description{
  ...
}
\usage{Chocolate}
\format{A data.frame containing 13 observations}
\source{C....}
\references{
  C. Alamprese, L. Datei, Q. Semeraro (2007).
  Optimization of processing parameters of a ball mill refiner for chocolate.
  \emph{Journal of Food Engineering}, 83(4), 629--636.
}
\keyword{datasets}
