\name{plot.harrington2}
\alias{plot.harrington2}
\title{Plot two sided Harrington type desirability}
\description{
  Plots a two sided desirability function of the Harrington type.
}
\usage{
\method{plot}{harrington2}(x, ...)
}
\arguments{
  \item{x}{rwo sided Harrington desriability function.}
  \item{\dots}{passed to plot.}
}
\author{
  Heike Trautmann \email{trautmann@statistik.uni-dortmund.de},
  Detlef Steuer \email{steuer@hsu-hamburg.de} and
  Olaf Mersmann \email{olafm@statistik.uni-dortmund.de}
}
\keyword{hplot}
