\name{geometricDI}
\alias{geometricDI}
\title{Geometric desirability index}
\description{
  TBD
}
\usage{
geometricDI(f, ..., weights)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{f, \dots}{desriability function}
  \item{weights}{vector of weights}
}
\value{
}
\author{
  Heike Trautmann \email{trautmann@statistik.uni-dortmund.de},
  Detlef Steuer \email{steuer@hsu-hamburg.de} and
  Olaf Mersmann \email{olafm@statistik.uni-dortmund.de}
}
\seealso{\code{\link{minimumDI}}}
\keyword{optimize}
\keyword{multivariate}
