\name{minimumDI}
\alias{minimumDI}
\title{Minimum desirability Index}
\description{
  TBD
}
\usage{
minimumDI(f, ...)
}
\arguments{
  \item{f,\dots}{desirability functions}
}
\author{
  Heike Trautmann \email{trautmann@statistik.uni-dortmund.de},
  Detlef Steuer \email{steuer@hsu-hamburg.de} and
  Olaf Mersmann \email{olafm@statistik.uni-dortmund.de}
}
\seealso{\code{\link{geometricDI}}}
\keyword{optimize}
\keyword{multivariate}
